+++ App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import Navigator from "./route/homeStack.js"
-import MenuLokalizacji from './screens/MenuLokalizacji';
-import PanelLokalizacjiEdytujUsun from './screens/PanelLokalizacjiEdytujUsun';
-import PanelLokalizacjiZapiszAnuluj from './screens/PanelLokalizacjiZapiszAnuluj';
+
 export default function App() {
-  return <MenuLokalizacji/> ;
+  return (
+    <Navigator/>
+  );
 }
