+++ route/homeStack.js
@@ -1,26 +1,15 @@
 import {createStackNavigator} from 'react-navigation-stack';
 import {createAppContainer} from 'react-navigation';
-import Home from '../screens/Home.js';
-import Login from '../screens/Login.js';
-import Dodawanie from '../screens/Dodawanie.js';
-import Report from '../screens/Report.js';
-import Test from '../screens/Test'
+import LocalisationMenu from '../screens/LocalisationMenu';
+import LocalisationEditDelete from '../screens/LocalisationEditDelete';
+import LocalisationSaveCancel from '../screens/LocalisationSaveCancel';
 
 
 const screens = {
     
     Test: {
-        screen: Test
-    },
-    Dodawanie: {
-        screen: Dodawanie
-    },
-    Home: {
-        screen: Home
-    },
-    Raport: {
-        screen: Report
-    },
+        screen: LocalisationMenu
+    }
 }
 
 const HomeStack = createStackNavigator(screens);
