+++ screens/LocalisationMenu.js
@@ -2,20 +2,26 @@ import React, {Component, useState} from "react";
 import {Button, StyleSheet, Text, View, Image, ImageBackground, TextInput} from "react-native";
 import {PasekNawigacyjnyMenuLokalizacji} from '../components/PasekNawigacyjny.js';
 import { InfoButton,PlusButton } from "../components/RoundButtons.js";
+import {Tray} from '../components/Trays';
+
 export default function App(){
   return (
-    <ImageBackground source={require('../assets/tlo_dodawanie.png')} style={styles.Tło}>
-    <View style={styles.Tło}>
+    <ImageBackground source={require('../assets/tlo_dodawanie.png')} style={{flex: 1}}>
+
+
       <View style={styles.paseknagorze}>
         <PasekNawigacyjnyMenuLokalizacji/>
       </View>
-      <View style={styles.button}>
+
+      <Tray composition='compact' spread='center'>
         <InfoButton/>
-      </View>
-      <View style={styles.button}>
+      </Tray>
+
+      <Tray composition='compact' spread='center'>
         <PlusButton/>
-      </View>
-    </View>
+      </Tray>
+
+
     </ImageBackground>
   );
 }
@@ -25,33 +31,5 @@ export default function App(){
       width: '100%',
       flexDirection: 'row',
       marginBottom: 50,
-    },
-
- 
-    button:
-    {
-     
-      alignItems: "center",
-      top:'10%',
-      height: "32%",
-    },
-
-    Tło: {
-      width: '100%',
-      height: '100%',      
-    },
-    tekst: {
-      position: "absolute",
-      left: '60%',
-      top:'1%',
-      color: 'black',
-      textAlign:'right',
-      fontSize: 25,
-      color:'white',
-      
-     
-
-     
-
     }
   })
